Seperate config dialogs into seperate components
