+++ Source/SelfService/Web/microservice/microserviceView/configuration/filesSection/validateFileDialog.tsx
@@ -5,7 +5,6 @@ import React from 'react';
 
 import { ConfirmDialog } from '@dolittle/design-system/atoms/ConfirmDialog/ConfirmDialog';
 
-import { GridSelectionModel } from '@mui/x-data-grid-pro';
 import { Box, Typography } from '@mui/material';
 
 const MAX_CONFIGMAP_ENTRY_SIZE = 3145728;
@@ -22,13 +21,6 @@ function formatBytes(bytes, decimals = 2) {
     return `${parseFloat((bytes / Math.pow(k, i)).toFixed(dm))} ${sizes[i]}`;
 };
 
-type DeleteConfigFileDialogProps = {
-    selectedDataRows: GridSelectionModel;
-    open: boolean;
-    setOpen: (open: boolean) => void;
-    handleDelete: () => Promise<void>;
-};
-
 type ValidateFileDialogProps = {
     invalid: {
         isOpen: boolean;
@@ -37,28 +29,6 @@ type ValidateFileDialogProps = {
     open: boolean;
     setOpen: () => void;
     handleValidate: () => void;
-};
-
-export const DeleteConfigFileDialog = ({ selectedDataRows, open, setOpen, handleDelete }: DeleteConfigFileDialogProps) => {
-    const hasManySelectedRows = selectedDataRows.length > 1;
-    const isPlural = hasManySelectedRows ? 'files' : 'file';
-
-    return (
-        <ConfirmDialog
-            id='delete-config-file-dialog'
-            title={`Delete configuration ${isPlural}`}
-            description={`Are you sure you want to delete ${hasManySelectedRows ? 'these' : 'this'} ${isPlural}?`}
-            cancelText='Cancel'
-            confirmText='Delete'
-            open={open}
-            handleCancel={() => setOpen(false)}
-            handleConfirm={handleDelete}
-        >
-            {selectedDataRows.map(file =>
-                <Typography key={file} variant='body2' sx={{ mt: 1.25 }}>{file}</Typography>
-            )}
-        </ConfirmDialog>
-    );
 };
 
 export const ValidateFileDialog = ({ invalid, open, setOpen, handleValidate }: ValidateFileDialogProps) => {
